==> PostProcessing/3_interpolate.ipynb <==
import os
import pandas as pd

def interpolate_missing_values(df):
    df.interpolate(method='linear', limit_direction='forward', axis=0, inplace=True)
    df.interpolate(method='linear', limit_direction='backward', axis=0, inplace=True)
    return df

def process_csv(folder_path):
    input_csv_file_path = os.path.join(folder_path, 'processed_step_2.csv')
    output_csv_file_path = os.path.join(folder_path, 'processed_step_3.csv')

    df = pd.read_csv(input_csv_file_path)

    last_valid_origin = df[['origin_x', 'origin_y']].last_valid_index()
    last_valid_goal = df[['goal_x', 'goal_y']].last_valid_index()

    df.loc[:last_valid_origin, ['origin_x', 'origin_y']] = interpolate_missing_values(df.loc[:last_valid_origin, ['origin_x', 'origin_y']])
    df.loc[:last_valid_goal, ['goal_x', 'goal_y']] = interpolate_missing_values(df.loc[:last_valid_goal, ['goal_x', 'goal_y']])

    df.to_csv(output_csv_file_path, index=False)

base_folder = r'.\NaoPickAndPlaceData'

for folder in os.listdir(base_folder):
    folder_path = os.path.join(base_folder, folder)
    if os.path.isdir(folder_path) and 'processed_step_2.csv' in os.listdir(folder_path):
        process_csv(folder_path)
