==> PostProcessing/5_is_grasp.ipynb <==
import os
import pandas as pd

def process_csv(folder_path):
    input_csv_file_path = os.path.join(folder_path, 'processed_step_4.csv') 
    output_csv_file_path = os.path.join(folder_path, 'processed_step_5.csv')

    df = pd.read_csv(input_csv_file_path)

    df['is_grasped'] = -1

    from_frame = 50
    baseline_x = df['origin_x'][:from_frame].mean()
    baseline_y = df['origin_y'][:from_frame].mean()

    threshold = 3

    for i in range(from_frame, len(df)):
        if abs(df.loc[i, 'origin_x'] - baseline_x) > threshold or abs(df.loc[i, 'origin_y'] - baseline_y) > threshold:
            df.loc[i, 'is_grasped'] = 1

    
    df.to_csv(output_csv_file_path, index=False)

base_folder = r'.\NaoPickAndPlaceData'
for folder in os.listdir(base_folder):
    folder_path = os.path.join(base_folder, folder)
    if os.path.isdir(folder_path) and 'processed_step_4.csv' in os.listdir(folder_path):
        process_csv(folder_path)
