==> PostProcessing/5_pushing.ipynb <==
import os
import pandas as pd

def process_csv(folder_path):
    input_csv_file_path = os.path.join(folder_path, 'processed_step_7.csv') 
    output_csv_file_path = os.path.join(folder_path, 'processed_step_10.csv')

    df = pd.read_csv(input_csv_file_path)

    df['direction'] = -1

    # Find the first and last valid ball_x values
    ball_x_values = df['origin_x'].dropna()
    first_ball_x = ball_x_values.iloc[0]
    last_ball_x = ball_x_values.iloc[-1]
    
    for i in range(0, len(df)):
        # Estimate the direction
        if last_ball_x > first_ball_x:
            df.loc[i, 'direction'] = 1  # Moving in a positive direction
        elif last_ball_x < first_ball_x:
            df.loc[i, 'direction'] = -1  # Moving in a negative direction
    
    df.to_csv(output_csv_file_path, index=False)

base_folder = r'.\NaoPushingData'
for folder in os.listdir(base_folder):
    folder_path = os.path.join(base_folder, folder)
    if os.path.isdir(folder_path) and 'processed_step_7.csv' in os.listdir(folder_path):
        process_csv(folder_path)
